getRNGState {httpuv}	R Documentation

Apply the value of .Random.seed to R's internal RNG state

Description

This function is needed in unusual cases where a C++ function calls an R function which sets the value of .Random.seed. This function should be called at the end of the R function to ensure that the new value .Random.seed is preserved. Otherwise, Rcpp may overwrite it with a previous value.

Usage

getRNGState()

[Package httpuv version 1.4.5 Index]