httpuv-package {httpuv}	R Documentation

HTTP and WebSocket server

Description

HTTP and WebSocket server

Details

Allows R code to listen for and interact with HTTP and WebSocket clients, so you can serve web traffic directly out of your R process. Implementation is based on libuv and http-parser.

This is a low-level library that provides little more than network I/O and implementations of the HTTP and WebSocket protocols. For an easy way to create web applications, try Shiny instead.

Author(s)

Joe Cheng joe@rstudio.com

See Also

startServer

Examples

## Not run: 
demo("echo", package="httpuv")

## End(Not run)

[Package httpuv version 1.4.5 Index]