interrupt {httpuv}	R Documentation

Interrupt httpuv runloop

Description

Interrupts the currently running httpuv runloop, meaning runServer or service will return control back to the caller and no further tasks will be processed until those methods are called again. Note that this may cause in-process uploads or downloads to be interrupted in mid-request.

Usage

interrupt()

[Package httpuv version 1.4.5 Index]