PipeServer {httpuv}	R Documentation

PipeServer class

Description

This class represents a server running one application that listens on a named pipe.

Methods

initialize(name, mask, app)

Create a new PipeServer object. app is an httpuv application object as described in startServer.

getName()

Return the value of name that was passed to initialize().

getMask()

Return the value of mask that was passed to initialize().

stop()

Stops a running server.

isRunning()

Returns TRUE if the server is currently running.

getStaticPaths()

Returns a list of staticPath objects for the server.

setStaticPath(..., .list = NULL)

Sets a static path for the current server. Each static path can be given as a named argument, or as an named item in .list. If there already exists a static path with the same name, it will be replaced.

removeStaticPath(path)

Removes a static path with the given name.

getStaticPathOptions()

Returns a list of default staticPathOptions for the current server. Each static path will use these options by default, but they can be overridden for each static path.

setStaticPathOption(..., .list = NULL)

Sets one or more static path options. Each option can be given as a named argument, or as a named item in .list.

Super class

httpuv::Server -> PipeServer

Methods

Public methods

PipeServer$new()

PipeServer$getName()

PipeServer$getMask()

Inherited methods

httpuv::Server$getStaticPathOptions()

httpuv::Server$getStaticPaths()

httpuv::Server$isRunning()

httpuv::Server$removeStaticPath()

httpuv::Server$setStaticPath()

httpuv::Server$setStaticPathOption()

httpuv::Server$stop()

Method new()

Usage

PipeServer$new(name, mask, app, quiet = FALSE)

Method getName()

Usage

PipeServer$getName()

Method getMask()

Usage

PipeServer$getMask()

See Also

Server and WebServer.

[Package httpuv version 1.6.3 Index]