Server class

The Server class is the parent class for WebServer and PipeServer. This class defines an interface and is not meant to be instantiated.

Methods

stop()

Stops a running server.

isRunning()

Returns TRUE if the server is currently running.

getStaticPaths()

Returns a list of staticPath objects for the server.

setStaticPath(..., .list = NULL)

Sets a static path for the current server. Each static path can be given as a named argument, or as an named item in .list. If there already exists a static path with the same name, it will be replaced.

removeStaticPath(path)

Removes a static path with the given name.

getStaticPathOptions()

Returns a list of default staticPathOptions for the current server. Each static path will use these options by default, but they can be overridden for each static path.

setStaticPathOption(..., .list = NULL)

Sets one or more static path options. Each option can be given as a named argument, or as a named item in .list.

See also

Methods

Public methods

Server$stop()

Server$isRunning()

Server$getStaticPaths()

Server$setStaticPath()

Server$removeStaticPath()

Server$getStaticPathOptions()

Server$setStaticPathOption()

Method stop()

Usage

Server$stop()

Method isRunning()

Usage

Server$isRunning()

Method getStaticPaths()

Usage

Server$getStaticPaths()

Method setStaticPath()

Usage

Server$setStaticPath(..., .list = NULL)

Method removeStaticPath()

Usage

Server$removeStaticPath(path)

Method getStaticPathOptions()

Usage

Server$getStaticPathOptions()

Method setStaticPathOption()

Usage

Server$setStaticPathOption(..., .list = NULL)