service {httpuv}	R Documentation

Process requests

Description

Process HTTP requests and WebSocket messages. Even if a server exists, no requests are serviced unless and until service is called.

Usage

service(timeoutMs = ifelse(interactive(), 100, 1000))

Arguments

timeoutMs
Approximate number of milliseconds to run before returning. If 0, then the function will continually process requests without returning unless an error occurs. If NA, performs a non-blocking run without waiting.

Details

Note that while service is waiting for a new request, the process is not interruptible using normal R means (Esc, Ctrl+C, etc.). If being interruptible is a requirement, then call service in a while loop with a very short but non-zero Sys.sleep during each iteration.

Examples

## Not run: 
while (TRUE) {
  service()
  Sys.sleep(0.001)
}

## End(Not run)

[Package httpuv version 1.3.6.2 Index]