stopDaemonizedServer {httpuv}	R Documentation

Stop a running daemonized server in Unix environments (deprecated)

Description

This function will be removed in a future release of httpuv. Instead, use stopServer.

Usage

stopDaemonizedServer(handle)

Arguments

handle
A handle that was previously returned from startServer or startPipeServer.

[Package httpuv version 1.4.5 Index]