stopDaemonizedServer {httpuv}	R Documentation

Stop a running daemonized server in Unix environments (deprecated)

Description

This function will be removed in a future release of httpuv. Instead, use stopServer.

Usage

stopDaemonizedServer(server)

Arguments

server
A server object that was previously returned from startServer or startPipeServer.

[Package httpuv version 1.6.15 Index]