stopDaemonizedServer {httpuv}	R Documentation

Stop a running daemonized server in Unix environments

Description

Given a handle that was returned from a previous invocation of startDaemonizedServer, closes all open connections for that server, removes listeners in the R event loop and unbinds the port. Be careful not to call stopDaemonizedServer more than once on a handle, as this will cause the R process to crash!

Usage

stopDaemonizedServer(server)

Arguments

server
A handle that was previously returned from startDaemonizedServer.

[Package httpuv version 1.3.6.2 Index]