stopServer {httpuv}	R Documentation

Stop a server

Description

Given a server object that was returned from a previous invocation of startServer or startPipeServer, this closes all open connections for that server and unbinds the port.

Usage

stopServer(server)

Arguments

server
A server object that was previously returned from startServer or startPipeServer.

See Also

stopAllServers to stop all servers.

[Package httpuv version 1.6.2 Index]